move simulator formatter into simulation namespace
