+++ lib/run/simulation_detailed.rb
@@ -10,8 +10,7 @@ module Run
       first_year_cash_flow_results = FirstYearCashFlow.new(app_config).calculate
       simulation_results = Simulation::Simulator.new(app_config).run
       evaluator_results = Simulation::SimulationEvaluator.new(simulation_results, app_config).evaluate
-      # TODO: Simulator Formatter should also move to simulator namespace and maybe call it ConsolePrinter
-      SimulatorFormatter.new(simulation_results, first_year_cash_flow_results, evaluator_results).print_all
+      Simulation::SimulatorFormatter.new(simulation_results, first_year_cash_flow_results, evaluator_results).print_all
     end
 
     private
